requirements.txt
asgiref==3.4.1
Django==3.0
djangorestframework==3.13.1
psycopg2-binary==2.8.6
pytz==2021.3
reverse==0.1.0
sqlparse==0.4.2
